Update principle-1.md
I've removed the CMS abbreviation throughout because in some instances it doesn't make sense to use the abbreviation,, and we should really be consistent. I think it reads fine without it, except for the introduction. I think we can probably do some more work here as 'case management systems' is written 5 times in the first few sentences and makes it a bit clunky. Just a suggestion though and happy to discuss.