Update principle-1.md
I thought when I first saw this that this section might need some more information on how to work with vendors to get this information and ensure that this happens? One of the big things that's come out of the most recent round of UR is a feeling of powerlessness with those LAs who aren't so techy. I saw that there's some info in the user needs section about working with vendors to achieve this, so maybe there could be a link to the user needs page in this sentence so they know immediately where to look next?