Update principle-5.md
With APIs a lot of the UR feedback was around the expense of them, even with the supplier they're currently with. We know from vendor work that they often have quite a lot of APIs on offer, but a few LAs said that the mainstream suppliers 'don't have APIs'. As such, I wonder if there's some guidance/SME content about APIs, finding them, getting the best deal for them etc?