Update user-needs.md
'A user is anyone who uses the system directly or uses outputs from the system' - I'd move this sentence to the very first paragraph under 'User needs'. It feels a bit out of place half way down the section. 
'Speak to other local authorities who have implemented new systems or developed their system, this could be done as part of a pre-market engagement exercise or something which you look to do ahead of developing an existing system' - have removed this sentence as the part about speaking with local authorities is repeated at the start of the section.